% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_predict.R
\name{custom_predict}
\alias{custom_predict}
\title{Custom Predict}
\usage{
custom_predict(object, ...)
}
\arguments{
\item{object}{a model object for which prediction is desired.}

\item{...}{additional arguments affecting the predictions produced.}
}
\description{
Custom Predict
}
