% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vetiverTest-package.R
\docType{package}
\name{vetiverTest-package}
\alias{vetiverTest}
\alias{vetiverTest-package}
\title{vetiverTest: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
